requirements.txt
Flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.25.0
python-dotenv==1.0.1
readability-lxml==0.8.1
lxml_html_clean==0.1.0
tldextract==5.3.0
gunicorn==22.0.0
google-search-results==2.4.2
httpx==0.27.0